refactor: /infra/place 파일구조를 /infra/client로 변경 #42
