+++ backend/src/main/java/com/dubu/backend/member/infra/client/NaverApiConfig.java
@@ -1,4 +1,4 @@
-package com.dubu.backend.member.infra.place;
+package com.dubu.backend.member.infra.client;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
